refactor: move test cards to a separate file
- these cards will also be used by TestsTab so it should not be in bootsTab.tsx
